Updated Final Report
Now only missing:
**Reflective reports will be added once QA'd**
4.1. Management Summary
4.2. Historical Account of Project
4.3. Performance of Team
4.4. Critical Evaluation of Team and Project
5.2.2. Appendix B2 – Failed Tests
5.4.2. Program Structure
5.4.3. Algorithms
5.4.4. Main Data Areas
5.4.5. Files
5.4.6. Interfaces
5.4.7. Suggestions for Improvements (needs a technical perspective)
5.4.9. Physical Limitations of Program
5.4.10. Rebuilding and Testing
+ ammendments from Bernies comments

Project Plan
Could have included a short overview of system implementation.
Timetable / Gantt chart: Too much detail, try to get it on one page.
Task allocation wasn’t a requirement for this hand in, but should be
useful to you anyway.
Use cases: OK.  A bit linear (although often are).  Not sure about the
server-server one, looks a bit basic.  Should have a description or
example scenario for each.

Design Specification
Section 3.2.1 Application in System seems to restate the requirements,
which is not really needed here.
The system parts listed along the top of the sequence diagrams should
correspond to classes or other  components of the system (e.g.
database), but don’t seem to.  As you are using a database, should
really include a database design (schema) i.e. tables with keys
etc.(Maybe state that we aren’t using a relational database?)
Any references for the servlets?